+++ src/1-symm/keccak.c.h
@@ -142,7 +142,7 @@ void glue(Keccak_InstName,_Permute)(void const *in, void *out)
             ptr[y*5+x] = htole(A[y][x]);
 }
 
-uintptr_t glue(i,Keccak_InstName)(int q){ return glue(c,Keccak_InstName)(q); }
+uintmax_t glue(i,Keccak_InstName)(int q){ return glue(c,Keccak_InstName)(q); }
 
 #undef Keccak_InstName
 #undef keccak_word_t
