+++ src/1-symm/rijndael.c
@@ -253,6 +253,6 @@ Define_AES_KeyExpansion(AES128_KeyExpansion,4,10);
 Define_AES_KeyExpansion(AES192_KeyExpansion,6,12);
 Define_AES_KeyExpansion(AES256_KeyExpansion,8,14);
 
-uintptr_t iAES128(int q) { return cAES128(q); }
-uintptr_t iAES192(int q) { return cAES192(q); }
-uintptr_t iAES256(int q) { return cAES256(q); }
+uintmax_t iAES128(int q) { return cAES128(q); }
+uintmax_t iAES192(int q) { return cAES192(q); }
+uintmax_t iAES256(int q) { return cAES256(q); }
