+++ src/2-xof/shake.c
@@ -34,5 +34,5 @@ void SHAKE_Read(shake_t *restrict x, void *restrict data, size_t len)
     Sponge_Read(&x->sponge, data, len);
 }
 
-uintptr_t iSHAKE128(int q){ return cSHAKE128(q); }
-uintptr_t iSHAKE256(int q){ return cSHAKE256(q); }
+uintmax_t iSHAKE128(int q){ return cSHAKE128(q); }
+uintmax_t iSHAKE256(int q){ return cSHAKE256(q); }
